backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.9
python-dotenv==1.0.1
cors==1.0.1
email-validator==2.2.0
pydantic==2.10.3
aiosmtplib==3.0.1